Added color picking
This allows the the player to click on an object of layer "map" and it will find the color of the pixel you clicked on and put that color on the answer block.

It will also print the value of how close the player was to the question color. Values closer to zero are better.